requirements.txt
selenium==3.141.0
Pillow==7.1.2